+++ trixnity-messenger-compose-view/src/iosMain/kotlin/de/connect2x/trixnity/messenger/compose/view/MultiMessengerViewController.kt
@@ -9,13 +9,13 @@ import com.arkivanov.decompose.DefaultComponentContext
 import com.arkivanov.essenty.lifecycle.ApplicationLifecycle
 import com.arkivanov.essenty.lifecycle.Lifecycle
 import de.connect2x.lognity.api.logger.Logger
-import de.connect2x.trixnity.messenger.compose.view.profiles.IntroductionOrProfile
+import de.connect2x.trixnity.messenger.MatrixMessengerSettingsHolder
+import de.connect2x.trixnity.messenger.MatrixMultiMessengerService
+import de.connect2x.trixnity.messenger.compose.view.profiles.Profiles
 import de.connect2x.trixnity.messenger.compose.view.profiles.ShowProfileCreation
 import de.connect2x.trixnity.messenger.compose.view.profiles.WithProfileSelection
 import de.connect2x.trixnity.messenger.compose.view.theme.IsFocusHighlighting
 import de.connect2x.trixnity.messenger.compose.view.theme.MessengerTheme
-import de.connect2x.trixnity.messenger.MatrixMessengerSettingsHolder
-import de.connect2x.trixnity.messenger.MatrixMultiMessengerService
 import de.connect2x.trixnity.messenger.delegateModule
 import de.connect2x.trixnity.messenger.multi.MatrixMultiMessengerConfiguration
 import de.connect2x.trixnity.messenger.uikit.WindowSceneDelegateProtocol
@@ -107,7 +107,7 @@ private fun MultiMessengerViewController(lifecycle: Lifecycle): UIViewController
                     IsFocusHighlighting provides false,
                 ) {
                     MessengerTheme {
-                        IntroductionOrProfile()
+                        Profiles()
                     }
                 }
             }