+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/RoomListRouter.kt
@@ -15,6 +15,8 @@ import de.connect2x.trixnity.messenger.util.popWhileSuspending
 import de.connect2x.trixnity.messenger.viewmodel.ViewModelContext
 import de.connect2x.trixnity.messenger.viewmodel.settings.AccountsViewModel
 import de.connect2x.trixnity.messenger.viewmodel.settings.AccountsViewModelFactory
+import de.connect2x.trixnity.messenger.viewmodel.settings.ProfilesSettingsViewModel
+import de.connect2x.trixnity.messenger.viewmodel.settings.ProfilesSettingsViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppInfoViewModel
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppInfoViewModelFactory
 import de.connect2x.trixnity.messenger.viewmodel.settings.AppearanceSettingsViewModel
@@ -150,10 +152,11 @@ class RoomListRouter(
 
             is Config.UserSettings -> Wrapper.UserSettings(
                 viewModelContext.get<UserSettingsViewModelFactory>().create(
-                    viewModelContext = viewModelContext.childContext("UserSettings", componentContext),
+                    viewModelContext = viewModelContext.childContext(name = "UserSettings", componentContext),
                     onCloseUserSettings = ::onCloseUserSettings,
                     onShowDeviceSettings = ::onShowDeviceSettings,
                     onShowAccounts = ::onShowAccounts,
+                    onShowProfilesSettings = ::onShowProfilesSettings,
                     onShowNotificationsSettings = ::onShowNotificationsSettings,
                     onShowPrivacySettings = ::onShowPrivacySettings,
                     onShowAppearanceSettings = ::onShowAppearanceSettings,
@@ -163,14 +166,14 @@ class RoomListRouter(
             is Config.DeviceSettings -> Wrapper.DeviceSettings(
                 viewModelContext.get<DeviceSettingsAllAccountsViewModelFactory>()
                     .create(
-                        viewModelContext = viewModelContext.childContext("DeviceSettings", componentContext),
+                        viewModelContext = viewModelContext.childContext(name = "DeviceSettings", componentContext),
                         onCloseDeviceSettings = ::onCloseDeviceSettings,
                     )
             )
 
             is Config.Accounts -> Wrapper.Accounts(
                 viewModelContext.get<AccountsViewModelFactory>().create(
-                    viewModelContext = viewModelContext.childContext("Accounts", componentContext),
+                    viewModelContext = viewModelContext.childContext(name = "Accounts", componentContext),
                     onCloseAccounts = ::onCloseAccounts,
                     onOpenAvatarCutter = onOpenAvatarCutter,
                     onShowAccountSetup = ::onShowAccountSetup,
@@ -179,17 +182,24 @@ class RoomListRouter(
                 )
             )
 
+            is Config.ProfilesSettings -> Wrapper.ProfilesSettings(
+                viewModelContext.get<ProfilesSettingsViewModelFactory>().create(
+                    viewModelContext = viewModelContext.childContext(name = "Profiles",componentContext),
+                    onCloseProfilesSettings = ::onCloseProfilesSettings,
+                )
+            )
+
             is Config.NotificationsSettings -> Wrapper.NotificationsSettings(
                 viewModelContext.get<NotificationSettingsAllAccountsViewModelFactory>()
                     .create(
-                        viewModelContext = viewModelContext.childContext("NotificationsSettings", componentContext),
+                        viewModelContext = viewModelContext.childContext(name = "NotificationsSettings", componentContext),
                         onBack = ::onCloseNotificationsSettings,
                     )
             )
 
             is Config.PrivacySettings -> Wrapper.PrivacySettings(
                 viewModelContext.get<PrivacySettingsAllAccountsViewModelFactory>().create(
-                    viewModelContext = viewModelContext.childContext("PrivacySettings", componentContext),
+                    viewModelContext = viewModelContext.childContext(name = "PrivacySettings", componentContext),
                     onClosePrivacySettings = ::onClosePrivacySettings,
                     onShowBlockedContactsSettings = ::onShowBlockedContactsSettings,
                 )
@@ -197,7 +207,7 @@ class RoomListRouter(
 
             is Config.AppearanceSettings -> Wrapper.AppearanceSettings(
                 viewModelContext.get<AppearanceSettingsViewModelFactory>().create(
-                    viewModelContext = viewModelContext.childContext("AppearanceSettings", componentContext),
+                    viewModelContext = viewModelContext.childContext(name = "AppearanceSettings", componentContext),
                     onCloseAppearanceSettings = ::onCloseAppearanceSettings,
                 )
             )
@@ -211,7 +221,7 @@ class RoomListRouter(
 
             is Config.AppInfo -> Wrapper.AppInfo(
                 viewModelContext.get<AppInfoViewModelFactory>().create(
-                    viewModelContext = viewModelContext.childContext("AppInfo", componentContext),
+                    viewModelContext = viewModelContext.childContext(name = "AppInfo", componentContext),
                     onCloseAppInfo = ::onCloseAppInfo,
                 )
             )
@@ -297,11 +307,19 @@ class RoomListRouter(
     }
 
     private fun onShowAccounts() {
-        log.debug { "show profile" }
+        log.debug { "show accounts" }
         navigation.launchPush(viewModelContext.coroutineScope, Config.Accounts)
     }
 
     private fun onCloseAccounts() {
+        log.debug { "close accounts" }
+        navigation.launchPop(viewModelContext.coroutineScope)
+    }
+    private fun onShowProfilesSettings() {
+        log.debug { "show profile" }
+        navigation.launchPush(viewModelContext.coroutineScope, Config.ProfilesSettings)
+    }
+    private fun onCloseProfilesSettings() {
         log.debug { "close profile" }
         navigation.launchPop(viewModelContext.coroutineScope)
     }
@@ -393,6 +411,9 @@ class RoomListRouter(
         data object Accounts : Config()
 
         @Serializable
+        data object ProfilesSettings : Config()
+
+        @Serializable
         data object NotificationsSettings : Config()
 
         @Serializable
@@ -419,6 +440,7 @@ class RoomListRouter(
         class UserSettings(val viewModel: UserSettingsViewModel) : Wrapper()
         class DeviceSettings(val viewModel: DeviceSettingsAllAccountsViewModel) : Wrapper()
         class Accounts(val viewModel: AccountsViewModel) : Wrapper()
+        class ProfilesSettings(val viewModel: ProfilesSettingsViewModel) : Wrapper()
         class NotificationsSettings(val viewModel: NotificationSettingsAllAccountsViewModel) : Wrapper()
         class PrivacySettings(val viewModel: PrivacySettingsAllAccountsViewModel) : Wrapper()
         class AppearanceSettings(val viewModel: AppearanceSettingsViewModel) : Wrapper()