+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - UI,SDK: Add option to disable redaction warning
+- SDK,UI Introduced profiles settings
 
 ### Changed